Refactored folder layout
